Split dispatcher and mcp/serve into focused submodules (#1425)
Mechanical no-behavior-change splits:

- crates/harn-vm/src/triggers/dispatcher/mod.rs (3.9k -> 2333 lines)
  - types.rs (398): Dispatcher + all data types + non-Dispatcher impls
  - state.rs (49): thread-local dispatch context + test signals
  - util.rs (415): control fns, time/event helpers, pub re-exported entrypoints
  - audit.rs (409): metadata builders for trust + lifecycle events
  - action_graph.rs (311): node/edge metadata + label helpers
  - vm_invoke.rs (169): VM callable invocation + timeout shim
  - mod.rs keeps the impl Dispatcher block intact (heart of dispatcher)

- crates/harn-cli/src/commands/mcp/serve.rs (3.7k lines) -> serve/
  - mod.rs (105): re-exports + LOG_STREAM_BINDINGS table + run entrypoint
  - types.rs (349): notification + state + request/response types
  - service.rs (152): McpOrchestratorService struct + ctor
  - protocol.rs (370): JSON-RPC handlers (impl block)
  - tools.rs (988): tool execution + task lifecycle (impl block)
  - resources.rs (548): resource/template handlers (impl block)
  - http.rs (758): HTTP transport, axum handlers, OAuth/legacy auth
  - stdio.rs (121): stdio transport
  - rpc_bridge.rs (61): RpcBridge impl
  - watchers.rs (350): list-change/log-topic notification watchers
  - util.rs (208): misc helpers

Each new module is under 1500 lines (largest: tools.rs at 988).
Public API preserved via pub use re-exports.
