+++ crates/harn-vm/src/triggers/dispatcher/mod.rs
@@ -1,32 +1,23 @@
-use std::cell::RefCell;
-use std::collections::{BTreeMap, BTreeSet};
+use std::collections::BTreeMap;
 use std::rc::Rc;
-use std::sync::atomic::{AtomicBool, AtomicU64, Ordering};
-use std::sync::{Arc, Mutex};
+use std::sync::atomic::Ordering;
+use std::sync::Arc;
 use std::time::{Duration, Instant};
 
 use futures::{pin_mut, StreamExt};
-use serde::{Deserialize, Serialize};
 use tokio::sync::broadcast;
-use tokio::sync::{Mutex as AsyncMutex, Notify};
+use tokio::sync::Mutex as AsyncMutex;
 use tracing::Instrument as _;
 
-use crate::event_log::{active_event_log, AnyEventLog, EventLog, LogError, LogEvent, Topic};
+use crate::event_log::{active_event_log, AnyEventLog, EventLog, LogEvent, Topic};
 use crate::llm::vm_value_to_json;
 use crate::orchestration::{
-    append_action_graph_update, RunActionGraphEdgeRecord, RunActionGraphNodeRecord,
-    RunObservabilityRecord, ACTION_GRAPH_EDGE_KIND_A2A_DISPATCH, ACTION_GRAPH_EDGE_KIND_DLQ_MOVE,
-    ACTION_GRAPH_EDGE_KIND_PREDICATE_GATE, ACTION_GRAPH_EDGE_KIND_REPLAY_CHAIN,
-    ACTION_GRAPH_EDGE_KIND_RETRY, ACTION_GRAPH_EDGE_KIND_TRIGGER_DISPATCH,
-    ACTION_GRAPH_NODE_KIND_A2A_HOP, ACTION_GRAPH_NODE_KIND_DISPATCH, ACTION_GRAPH_NODE_KIND_DLQ,
+    RunActionGraphEdgeRecord, RunActionGraphNodeRecord, ACTION_GRAPH_EDGE_KIND_DLQ_MOVE,
+    ACTION_GRAPH_EDGE_KIND_REPLAY_CHAIN, ACTION_GRAPH_EDGE_KIND_RETRY,
+    ACTION_GRAPH_EDGE_KIND_TRIGGER_DISPATCH, ACTION_GRAPH_NODE_KIND_DLQ,
     ACTION_GRAPH_NODE_KIND_PREDICATE, ACTION_GRAPH_NODE_KIND_RETRY, ACTION_GRAPH_NODE_KIND_TRIGGER,
-    ACTION_GRAPH_NODE_KIND_WORKER_ENQUEUE,
 };
-use crate::stdlib::json_to_vm_value;
-use crate::trust_graph::{
-    append_trust_record, policy_for_autonomy_tier, AutonomyTier, TrustOutcome, TrustRecord,
-};
-use crate::value::{error_to_category, ErrorCategory, VmError, VmValue};
+use crate::trust_graph::{AutonomyTier, TrustOutcome};
 use crate::vm::Vm;
 
 use self::uri::DispatchUri;
@@ -41,450 +32,62 @@ use super::{
 };
 use circuits::{
     destination_circuit_key, dlq_node_metadata, DestinationCircuitProbe,
-    DestinationCircuitRegistry, DESTINATION_CIRCUIT_FAILURE_THRESHOLD,
+    DESTINATION_CIRCUIT_FAILURE_THRESHOLD,
 };
-use flow_control::{BatchDecision, ConcurrencyPermit, FlowControlManager};
+use flow_control::BatchDecision;
 use predicate_eval::predicate_node_metadata;
 
+mod action_graph;
+mod audit;
 mod circuits;
 mod flow_control;
 mod predicate_eval;
 pub mod retry;
+mod state;
+mod types;
 pub mod uri;
+mod util;
+mod vm_invoke;
 
 pub use retry::{RetryPolicy, TriggerRetryConfig, DEFAULT_MAX_ATTEMPTS};
+pub use types::{
+    DispatchAttemptRecord, DispatchCancelRequest, DispatchError, DispatchOutcome, DispatchStatus,
+    Dispatcher, DispatcherDrainReport, DispatcherStatsSnapshot, DlqEntry, InboxEnvelope,
+};
+pub use util::{
+    append_dispatch_cancel_request, clear_dispatcher_state, enqueue_trigger_event,
+    snapshot_dispatcher_stats,
+};
+
+pub(crate) use state::{
+    current_dispatch_context, current_dispatch_is_replay, current_dispatch_wait_lease,
+};
+pub(crate) use types::{DispatchContext, DispatchWaitLease};
+
+use action_graph::{
+    dispatch_entry_edge_kind, dispatch_error_label, dispatch_error_metadata, dispatch_node_id,
+    dispatch_node_kind, dispatch_node_label, dispatch_node_metadata, dispatch_success_metadata,
+    dispatch_success_outcome, dispatch_target_agent, retry_node_metadata,
+    trigger_event_persona_metadata, trigger_node_metadata,
+};
+#[cfg(test)]
+use state::install_test_inbox_dequeued_signal;
+use state::{notify_test_inbox_dequeued, ACTIVE_DISPATCHER_STATE, ACTIVE_DISPATCH_IS_REPLAY};
+use types::{
+    AcquiredFlowControl, ConcurrencyLease, DispatchSkipStage, DispatcherRuntimeState,
+    FlowControlOutcome, SingletonLease, DEFAULT_AUTONOMY_BUDGET_REVIEWER,
+};
+use util::{
+    accepted_at_ms, binding_key_from_parts, build_batched_event, cancelled_dispatch_outcome,
+    current_unix_ms, decrement_in_flight, decrement_retry_queue_depth, dispatch_cancel_requested,
+    dispatch_result_status, duration_between_ms, event_headers, json_value_to_gate,
+    maybe_fail_before_outbox, now_rfc3339, now_unix_ms, queue_appended_at_ms, recv_cancel,
+    sleep_or_cancel_or_request, tenant_id, topic_for_event, unix_ms, worker_queue_priority,
+};
 
 pub const TRIGGER_ACCEPTED_AT_MS_HEADER: &str = "harn_trigger_accepted_at_ms";
 pub const TRIGGER_NORMALIZED_AT_MS_HEADER: &str = "harn_trigger_normalized_at_ms";
 pub const TRIGGER_QUEUE_APPENDED_AT_MS_HEADER: &str = "harn_trigger_queue_appended_at_ms";
-
-thread_local! {
-    static ACTIVE_DISPATCHER_STATE: RefCell<Option<Arc<DispatcherRuntimeState>>> = const { RefCell::new(None) };
-    static ACTIVE_DISPATCH_CONTEXT: RefCell<Option<DispatchContext>> = const { RefCell::new(None) };
-    static ACTIVE_DISPATCH_WAIT_LEASE: RefCell<Option<DispatchWaitLease>> = const { RefCell::new(None) };
-    #[cfg(test)]
-    static TEST_INBOX_DEQUEUED_SIGNAL: RefCell<Option<tokio::sync::oneshot::Sender<()>>> = const { RefCell::new(None) };
-}
-
-tokio::task_local! {
-    static ACTIVE_DISPATCH_IS_REPLAY: bool;
-}
-
-#[derive(Clone, Debug)]
-pub(crate) struct DispatchContext {
-    pub trigger_event: TriggerEvent,
-    pub replay_of_event_id: Option<String>,
-    pub binding_id: String,
-    pub binding_version: u32,
-    pub agent_id: String,
-    pub action: String,
-    pub autonomy_tier: AutonomyTier,
-}
-
-struct DispatchExecutionPolicyGuard;
-
-impl Drop for DispatchExecutionPolicyGuard {
-    fn drop(&mut self) {
-        crate::orchestration::pop_execution_policy();
-    }
-}
-
-const DEFAULT_AUTONOMY_BUDGET_REVIEWER: &str = "operator";
-
-pub(crate) fn current_dispatch_context() -> Option<DispatchContext> {
-    ACTIVE_DISPATCH_CONTEXT.with(|slot| slot.borrow().clone())
-}
-
-pub(crate) fn current_dispatch_is_replay() -> bool {
-    ACTIVE_DISPATCH_IS_REPLAY
-        .try_with(|is_replay| *is_replay)
-        .unwrap_or(false)
-}
-
-pub(crate) fn current_dispatch_wait_lease() -> Option<DispatchWaitLease> {
-    ACTIVE_DISPATCH_WAIT_LEASE.with(|slot| slot.borrow().clone())
-}
-
-#[derive(Clone)]
-pub struct Dispatcher {
-    base_vm: Rc<Vm>,
-    event_log: Arc<AnyEventLog>,
-    cancel_tx: broadcast::Sender<()>,
-    state: Arc<DispatcherRuntimeState>,
-    metrics: Option<Arc<crate::MetricsRegistry>>,
-    a2a_client: Arc<dyn crate::a2a::A2aClient>,
-}
-
-#[derive(Debug)]
-struct DispatcherRuntimeState {
-    in_flight: AtomicU64,
-    retry_queue_depth: AtomicU64,
-    dlq: Mutex<Vec<DlqEntry>>,
-    cancel_tokens: Mutex<Vec<Arc<std::sync::atomic::AtomicBool>>>,
-    shutting_down: std::sync::atomic::AtomicBool,
-    idle_notify: Notify,
-    flow_control: FlowControlManager,
-    destination_circuits: DestinationCircuitRegistry,
-}
-
-impl DispatcherRuntimeState {
-    fn new(event_log: Arc<AnyEventLog>) -> Self {
-        Self {
-            in_flight: AtomicU64::new(0),
-            retry_queue_depth: AtomicU64::new(0),
-            dlq: Mutex::new(Vec::new()),
-            cancel_tokens: Mutex::new(Vec::new()),
-            shutting_down: std::sync::atomic::AtomicBool::new(false),
-            idle_notify: Notify::new(),
-            flow_control: FlowControlManager::new(event_log),
-            destination_circuits: DestinationCircuitRegistry::default(),
-        }
-    }
-}
-
-#[derive(Clone, Debug, Default, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(default)]
-pub struct DispatcherStatsSnapshot {
-    pub in_flight: u64,
-    pub retry_queue_depth: u64,
-    pub dlq_depth: u64,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum DispatchStatus {
-    Succeeded,
-    Failed,
-    Dlq,
-    Skipped,
-    Waiting,
-    Cancelled,
-}
-
-impl DispatchStatus {
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            Self::Succeeded => "succeeded",
-            Self::Failed => "failed",
-            Self::Dlq => "dlq",
-            Self::Skipped => "skipped",
-            Self::Waiting => "waiting",
-            Self::Cancelled => "cancelled",
-        }
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
-#[serde(default)]
-pub struct DispatchOutcome {
-    pub trigger_id: String,
-    pub binding_key: String,
-    pub event_id: String,
-    pub attempt_count: u32,
-    pub status: DispatchStatus,
-    pub handler_kind: String,
-    pub target_uri: String,
-    pub replay_of_event_id: Option<String>,
-    pub result: Option<serde_json::Value>,
-    pub error: Option<String>,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct InboxEnvelope {
-    pub trigger_id: Option<String>,
-    pub binding_version: Option<u32>,
-    pub event: TriggerEvent,
-}
-
-#[derive(Clone, Debug, Default, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(default)]
-pub struct DispatcherDrainReport {
-    pub drained: bool,
-    pub in_flight: u64,
-    pub retry_queue_depth: u64,
-    pub dlq_depth: u64,
-}
-
-impl Default for DispatchOutcome {
-    fn default() -> Self {
-        Self {
-            trigger_id: String::new(),
-            binding_key: String::new(),
-            event_id: String::new(),
-            attempt_count: 0,
-            status: DispatchStatus::Failed,
-            handler_kind: String::new(),
-            target_uri: String::new(),
-            replay_of_event_id: None,
-            result: None,
-            error: None,
-        }
-    }
-}
-
-#[derive(Clone, Debug, Default, PartialEq, Serialize, Deserialize)]
-#[serde(default)]
-pub struct DispatchAttemptRecord {
-    pub trigger_id: String,
-    pub binding_key: String,
-    pub event_id: String,
-    pub attempt: u32,
-    pub handler_kind: String,
-    pub started_at: String,
-    pub completed_at: String,
-    pub outcome: String,
-    pub error_msg: Option<String>,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize)]
-pub struct DispatchCancelRequest {
-    pub binding_key: String,
-    pub event_id: String,
-    #[serde(with = "time::serde::rfc3339")]
-    pub requested_at: time::OffsetDateTime,
-    #[serde(default)]
-    pub requested_by: Option<String>,
-    #[serde(default)]
-    pub audit_id: Option<String>,
-}
-
-#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
-pub struct DlqEntry {
-    pub trigger_id: String,
-    pub binding_key: String,
-    pub event: TriggerEvent,
-    pub attempt_count: u32,
-    pub final_error: String,
-    #[serde(default = "default_dlq_error_class")]
-    pub error_class: String,
-    pub attempts: Vec<DispatchAttemptRecord>,
-}
-
-fn default_dlq_error_class() -> String {
-    "unknown".to_string()
-}
-
-#[derive(Clone, Debug)]
-struct SingletonLease {
-    gate: String,
-    held: bool,
-}
-
-#[derive(Clone, Debug)]
-struct ConcurrencyLease {
-    gate: String,
-    max: u32,
-    priority_rank: usize,
-    permit: Option<ConcurrencyPermit>,
-}
-
-#[derive(Default, Debug)]
-struct AcquiredFlowControl {
-    singleton: Option<SingletonLease>,
-    concurrency: Option<ConcurrencyLease>,
-}
-
-#[derive(Clone)]
-pub(crate) struct DispatchWaitLease {
-    state: Arc<DispatcherRuntimeState>,
-    acquired: Arc<AsyncMutex<AcquiredFlowControl>>,
-    suspended: Arc<AtomicBool>,
-}
-
-impl DispatchWaitLease {
-    fn new(
-        state: Arc<DispatcherRuntimeState>,
-        acquired: Arc<AsyncMutex<AcquiredFlowControl>>,
-    ) -> Self {
-        Self {
-            state,
-            acquired,
-            suspended: Arc::new(AtomicBool::new(false)),
-        }
-    }
-
-    pub(crate) async fn suspend(&self) -> Result<(), DispatchError> {
-        if self.suspended.swap(true, Ordering::SeqCst) {
-            return Ok(());
-        }
-        let (singleton_gate, concurrency_permit) = {
-            let mut acquired = self.acquired.lock().await;
-            let singleton_gate = acquired.singleton.as_mut().and_then(|lease| {
-                if lease.held {
-                    lease.held = false;
-                    Some(lease.gate.clone())
-                } else {
-                    None
-                }
-            });
-            let concurrency_permit = acquired
-                .concurrency
-                .as_mut()
-                .and_then(|lease| lease.permit.take());
-            (singleton_gate, concurrency_permit)
-        };
-
-        if let Some(gate) = singleton_gate {
-            self.state
-                .flow_control
-                .release_singleton(&gate)
-                .await
-                .map_err(DispatchError::from)?;
-        }
-        if let Some(permit) = concurrency_permit {
-            self.state
-                .flow_control
-                .release_concurrency(permit)
-                .await
-                .map_err(DispatchError::from)?;
-        }
-        Ok(())
-    }
-
-    pub(crate) async fn resume(&self) -> Result<(), DispatchError> {
-        if !self.suspended.swap(false, Ordering::SeqCst) {
-            return Ok(());
-        }
-
-        let singleton_gate = {
-            let acquired = self.acquired.lock().await;
-            acquired.singleton.as_ref().and_then(|lease| {
-                if lease.held {
-                    None
-                } else {
-                    Some(lease.gate.clone())
-                }
-            })
-        };
-        if let Some(gate) = singleton_gate {
-            self.state
-                .flow_control
-                .acquire_singleton(&gate)
-                .await
-                .map_err(DispatchError::from)?;
-            let mut acquired = self.acquired.lock().await;
-            if let Some(lease) = acquired.singleton.as_mut() {
-                lease.held = true;
-            }
-        }
-
-        let concurrency_spec = {
-            let acquired = self.acquired.lock().await;
-            acquired.concurrency.as_ref().and_then(|lease| {
-                if lease.permit.is_some() {
-                    None
-                } else {
-                    Some((lease.gate.clone(), lease.max, lease.priority_rank))
-                }
-            })
-        };
-        if let Some((gate, max, priority_rank)) = concurrency_spec {
-            let permit = self
-                .state
-                .flow_control
-                .acquire_concurrency(&gate, max, priority_rank)
-                .await
-                .map_err(DispatchError::from)?;
-            let mut acquired = self.acquired.lock().await;
-            if let Some(lease) = acquired.concurrency.as_mut() {
-                lease.permit = Some(permit);
-            }
-        }
-        Ok(())
-    }
-}
-
-enum FlowControlOutcome {
-    Dispatch {
-        event: Box<TriggerEvent>,
-        acquired: AcquiredFlowControl,
-    },
-    Skip {
-        reason: String,
-    },
-}
-
-#[derive(Clone, Debug)]
-enum DispatchSkipStage {
-    Predicate,
-    FlowControl,
-}
-
-#[derive(Debug)]
-pub enum DispatchError {
-    EventLog(String),
-    Registry(String),
-    Serde(String),
-    Local(String),
-    A2a(String),
-    Denied(String),
-    Timeout(String),
-    Waiting(String),
-    Cancelled(String),
-    NotImplemented(String),
-}
-
-impl std::fmt::Display for DispatchError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::EventLog(message)
-            | Self::Registry(message)
-            | Self::Serde(message)
-            | Self::Local(message)
-            | Self::A2a(message)
-            | Self::Denied(message)
-            | Self::Timeout(message)
-            | Self::Waiting(message)
-            | Self::Cancelled(message)
-            | Self::NotImplemented(message) => f.write_str(message),
-        }
-    }
-}
-
-impl std::error::Error for DispatchError {}
-
-impl DispatchError {
-    fn retryable(&self) -> bool {
-        !matches!(
-            self,
-            Self::Cancelled(_) | Self::Denied(_) | Self::NotImplemented(_) | Self::Waiting(_)
-        )
-    }
-}
-
-impl DispatchSkipStage {
-    fn as_str(&self) -> &'static str {
-        match self {
-            Self::Predicate => "predicate",
-            Self::FlowControl => "flow_control",
-        }
-    }
-}
-
-impl From<LogError> for DispatchError {
-    fn from(value: LogError) -> Self {
-        Self::EventLog(value.to_string())
-    }
-}
-
-pub async fn append_dispatch_cancel_request(
-    event_log: &Arc<AnyEventLog>,
-    request: &DispatchCancelRequest,
-) -> Result<u64, DispatchError> {
-    let topic = Topic::new(TRIGGER_CANCEL_REQUESTS_TOPIC)
-        .expect("static trigger cancel topic should always be valid");
-    event_log
-        .append(
-            &topic,
-            LogEvent::new(
-                "dispatch_cancel_requested",
-                serde_json::to_value(request)
-                    .map_err(|error| DispatchError::Serde(error.to_string()))?,
-            ),
-        )
-        .await
-        .map_err(DispatchError::from)
-}
 
 impl Dispatcher {
     pub fn event_log_handle(&self) -> Arc<AnyEventLog> {
@@ -2724,1202 +2327,6 @@ impl Dispatcher {
             .await?;
         Ok(json_value_to_gate(&vm_value_to_json(&value)))
     }
-
-    #[allow(clippy::too_many_arguments)]
-    async fn invoke_vm_callable(
-        &self,
-        closure: &crate::value::VmClosure,
-        binding_key: &str,
-        event: &TriggerEvent,
-        replay_of_event_id: Option<&String>,
-        agent_id: &str,
-        action: &str,
-        autonomy_tier: AutonomyTier,
-        wait_lease: Option<DispatchWaitLease>,
-        cancel_rx: &mut broadcast::Receiver<()>,
-    ) -> Result<VmValue, DispatchError> {
-        let mut vm = self.base_vm.child_vm();
-        let cancel_token = Arc::new(std::sync::atomic::AtomicBool::new(false));
-        if self.state.shutting_down.load(Ordering::SeqCst) {
-            cancel_token.store(true, Ordering::SeqCst);
-        }
-        self.state
-            .cancel_tokens
-            .lock()
-            .expect("dispatcher cancel tokens poisoned")
-            .push(cancel_token.clone());
-        vm.install_cancel_token(cancel_token.clone());
-        let arg = event_to_handler_value(event)?;
-        let args = [arg];
-        let tier_policy = policy_for_autonomy_tier(autonomy_tier);
-        let effective_policy = match crate::orchestration::current_execution_policy() {
-            Some(parent) => parent
-                .intersect(&tier_policy)
-                .map_err(|error| DispatchError::Local(error.to_string()))?,
-            None => tier_policy,
-        };
-        crate::orchestration::push_execution_policy(effective_policy);
-        let _policy_guard = DispatchExecutionPolicyGuard;
-        let future = vm.call_closure_pub(closure, &args);
-        pin_mut!(future);
-        let (binding_id, binding_version) = split_binding_key(binding_key);
-        let prior_context = ACTIVE_DISPATCH_CONTEXT.with(|slot| {
-            slot.borrow_mut().replace(DispatchContext {
-                trigger_event: event.clone(),
-                replay_of_event_id: replay_of_event_id.cloned(),
-                binding_id,
-                binding_version,
-                agent_id: agent_id.to_string(),
-                action: action.to_string(),
-                autonomy_tier,
-            })
-        });
-        let prior_wait_lease = ACTIVE_DISPATCH_WAIT_LEASE
-            .with(|slot| std::mem::replace(&mut *slot.borrow_mut(), wait_lease));
-        let prior_hitl_state = crate::stdlib::hitl::take_hitl_state();
-        crate::stdlib::hitl::reset_hitl_state();
-        let mut poll = tokio::time::interval(Duration::from_millis(100));
-        let result = loop {
-            tokio::select! {
-                result = &mut future => break result,
-                _ = recv_cancel(cancel_rx) => {
-                    cancel_token.store(true, Ordering::SeqCst);
-                }
-                _ = poll.tick() => {
-                    if dispatch_cancel_requested(
-                        &self.event_log,
-                        binding_key,
-                        &event.id.0,
-                        replay_of_event_id,
-                    )
-                    .await? {
-                        cancel_token.store(true, Ordering::SeqCst);
-                    }
-                }
-            }
-        };
-        ACTIVE_DISPATCH_CONTEXT.with(|slot| {
-            *slot.borrow_mut() = prior_context;
-        });
-        ACTIVE_DISPATCH_WAIT_LEASE.with(|slot| {
-            *slot.borrow_mut() = prior_wait_lease;
-        });
-        crate::stdlib::hitl::restore_hitl_state(prior_hitl_state);
-        {
-            let mut tokens = self
-                .state
-                .cancel_tokens
-                .lock()
-                .expect("dispatcher cancel tokens poisoned");
-            tokens.retain(|token| !Arc::ptr_eq(token, &cancel_token));
-        }
-
-        if cancel_token.load(Ordering::SeqCst) {
-            if dispatch_cancel_requested(
-                &self.event_log,
-                binding_key,
-                &event.id.0,
-                replay_of_event_id,
-            )
-            .await?
-            {
-                Err(DispatchError::Cancelled(
-                    "trigger cancel request cancelled local handler".to_string(),
-                ))
-            } else {
-                Err(DispatchError::Cancelled(
-                    "dispatcher shutdown cancelled local handler".to_string(),
-                ))
-            }
-        } else {
-            result.map_err(dispatch_error_from_vm_error)
-        }
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    async fn invoke_vm_callable_with_timeout(
-        &self,
-        closure: &crate::value::VmClosure,
-        binding_key: &str,
-        event: &TriggerEvent,
-        replay_of_event_id: Option<&String>,
-        agent_id: &str,
-        action: &str,
-        autonomy_tier: AutonomyTier,
-        cancel_rx: &mut broadcast::Receiver<()>,
-        timeout: Option<Duration>,
-    ) -> Result<VmValue, DispatchError> {
-        let future = self.invoke_vm_callable(
-            closure,
-            binding_key,
-            event,
-            replay_of_event_id,
-            agent_id,
-            action,
-            autonomy_tier,
-            None,
-            cancel_rx,
-        );
-        pin_mut!(future);
-        if let Some(timeout) = timeout {
-            match tokio::time::timeout(timeout, future).await {
-                Ok(result) => result,
-                Err(_) => Err(DispatchError::Local(
-                    "predicate evaluation timed out".to_string(),
-                )),
-            }
-        } else {
-            future.await
-        }
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    async fn append_dispatch_trust_record(
-        &self,
-        binding: &TriggerBinding,
-        route: &DispatchUri,
-        event: &TriggerEvent,
-        replay_of_event_id: Option<&String>,
-        autonomy_tier: AutonomyTier,
-        outcome: TrustOutcome,
-        terminal_status: &str,
-        attempt_count: u32,
-        error: Option<String>,
-    ) -> Result<(), DispatchError> {
-        let mut record = TrustRecord::new(
-            binding.id.as_str().to_string(),
-            format!("{}.{}", event.provider.as_str(), event.kind),
-            None,
-            outcome,
-            event.trace_id.0.clone(),
-            autonomy_tier,
-        );
-        record.metadata.insert(
-            "binding_key".to_string(),
-            serde_json::json!(binding.binding_key()),
-        );
-        record.metadata.insert(
-            "binding_version".to_string(),
-            serde_json::json!(binding.version),
-        );
-        record.metadata.insert(
-            "provider".to_string(),
-            serde_json::json!(event.provider.as_str()),
-        );
-        record
-            .metadata
-            .insert("event_kind".to_string(), serde_json::json!(event.kind));
-        record
-            .metadata
-            .insert("handler_kind".to_string(), serde_json::json!(route.kind()));
-        record.metadata.insert(
-            "target_uri".to_string(),
-            serde_json::json!(route.target_uri()),
-        );
-        record.metadata.insert(
-            "terminal_status".to_string(),
-            serde_json::json!(terminal_status),
-        );
-        record.metadata.insert(
-            "attempt_count".to_string(),
-            serde_json::json!(attempt_count),
-        );
-        if let Some(replay_of_event_id) = replay_of_event_id {
-            record.metadata.insert(
-                "replay_of_event_id".to_string(),
-                serde_json::json!(replay_of_event_id),
-            );
-        }
-        if let Some(error) = error {
-            record
-                .metadata
-                .insert("error".to_string(), serde_json::json!(error));
-        }
-        append_trust_record(&self.event_log, &record)
-            .await
-            .map(|_| ())
-            .map_err(DispatchError::from)
-    }
-
-    async fn append_attempt_record(
-        &self,
-        event: &TriggerEvent,
-        binding: &TriggerBinding,
-        attempt: &DispatchAttemptRecord,
-        replay_of_event_id: Option<&String>,
-    ) -> Result<(), DispatchError> {
-        self.append_topic_event(
-            TRIGGER_ATTEMPTS_TOPIC,
-            "attempt_recorded",
-            event,
-            Some(binding),
-            Some(attempt.attempt),
-            serde_json::to_value(attempt)
-                .map_err(|error| DispatchError::Serde(error.to_string()))?,
-            replay_of_event_id,
-        )
-        .await
-    }
-
-    async fn append_lifecycle_event(
-        &self,
-        kind: &str,
-        event: &TriggerEvent,
-        binding: &TriggerBinding,
-        payload: serde_json::Value,
-        replay_of_event_id: Option<&String>,
-    ) -> Result<(), DispatchError> {
-        self.append_topic_event(
-            TRIGGERS_LIFECYCLE_TOPIC,
-            kind,
-            event,
-            Some(binding),
-            None,
-            payload,
-            replay_of_event_id,
-        )
-        .await
-    }
-
-    async fn append_autonomy_budget_approval_request(
-        &self,
-        binding: &TriggerBinding,
-        route: &DispatchUri,
-        event: &TriggerEvent,
-        replay_of_event_id: Option<&String>,
-        reason: &str,
-    ) -> Result<String, DispatchError> {
-        let reviewers = vec![DEFAULT_AUTONOMY_BUDGET_REVIEWER.to_string()];
-        let detail = serde_json::json!({
-            "trigger_id": binding.id.as_str(),
-            "binding_key": binding.binding_key(),
-            "event_id": event.id.0,
-            "reason": reason,
-            "from_tier": AutonomyTier::ActAuto.as_str(),
-            "requested_tier": AutonomyTier::ActWithApproval.as_str(),
-            "handler_kind": route.kind(),
-            "target_uri": route.target_uri(),
-            "max_autonomous_decisions_per_hour": binding.max_autonomous_decisions_per_hour,
-            "max_autonomous_decisions_per_day": binding.max_autonomous_decisions_per_day,
-            "autonomous_decisions_hour": binding.metrics.autonomous_decisions_hour.load(Ordering::Relaxed),
-            "autonomous_decisions_today": binding.metrics.autonomous_decisions_today.load(Ordering::Relaxed),
-            "replay_of_event_id": replay_of_event_id,
-        });
-        let request_id = crate::stdlib::hitl::append_approval_request_on(
-            &self.event_log,
-            binding.id.as_str().to_string(),
-            event.trace_id.0.clone(),
-            format!(
-                "approve autonomous dispatch for trigger '{}' after {}",
-                binding.id.as_str(),
-                reason
-            ),
-            detail.clone(),
-            reviewers.clone(),
-        )
-        .await
-        .map_err(dispatch_error_from_vm_error)?;
-        self.append_lifecycle_event(
-            "autonomy.budget_exceeded",
-            event,
-            binding,
-            serde_json::json!({
-                "trigger_id": binding.id.as_str(),
-                "event_id": event.id.0,
-                "reason": reason,
-                "request_id": request_id,
-                "reviewers": reviewers,
-                "from_tier": AutonomyTier::ActAuto.as_str(),
-                "requested_tier": AutonomyTier::ActWithApproval.as_str(),
-                "replay_of_event_id": replay_of_event_id,
-            }),
-            replay_of_event_id,
-        )
-        .await?;
-        Ok(request_id)
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    async fn emit_autonomy_budget_approval_action_graph(
-        &self,
-        binding: &TriggerBinding,
-        route: &DispatchUri,
-        event: &TriggerEvent,
-        source_node_id: &str,
-        replay_of_event_id: Option<&String>,
-        reason: &str,
-        request_id: &str,
-    ) -> Result<(), DispatchError> {
-        let approval_node_id = format!("approval:{}:{}", binding.binding_key(), event.id.0);
-        self.emit_action_graph(
-            event,
-            vec![RunActionGraphNodeRecord {
-                id: approval_node_id.clone(),
-                label: format!("approval required: {reason}"),
-                kind: "approval".to_string(),
-                status: "waiting".to_string(),
-                outcome: "request_approval".to_string(),
-                trace_id: Some(event.trace_id.0.clone()),
-                stage_id: None,
-                node_id: None,
-                worker_id: None,
-                run_id: None,
-                run_path: None,
-                metadata: BTreeMap::from([
-                    (
-                        "trigger_id".to_string(),
-                        serde_json::json!(binding.id.as_str()),
-                    ),
-                    (
-                        "binding_key".to_string(),
-                        serde_json::json!(binding.binding_key()),
-                    ),
-                    ("event_id".to_string(), serde_json::json!(event.id.0)),
-                    ("reason".to_string(), serde_json::json!(reason)),
-                    ("request_id".to_string(), serde_json::json!(request_id)),
-                    (
-                        "reviewers".to_string(),
-                        serde_json::json!([DEFAULT_AUTONOMY_BUDGET_REVIEWER]),
-                    ),
-                    ("handler_kind".to_string(), serde_json::json!(route.kind())),
-                    (
-                        "target_uri".to_string(),
-                        serde_json::json!(route.target_uri()),
-                    ),
-                    (
-                        "from_tier".to_string(),
-                        serde_json::json!(AutonomyTier::ActAuto.as_str()),
-                    ),
-                    (
-                        "requested_tier".to_string(),
-                        serde_json::json!(AutonomyTier::ActWithApproval.as_str()),
-                    ),
-                    (
-                        "replay_of_event_id".to_string(),
-                        serde_json::json!(replay_of_event_id),
-                    ),
-                ]),
-            }],
-            vec![RunActionGraphEdgeRecord {
-                from_id: source_node_id.to_string(),
-                to_id: approval_node_id,
-                kind: "approval_gate".to_string(),
-                label: Some("autonomy budget".to_string()),
-            }],
-            serde_json::json!({
-                "source": "dispatcher",
-                "trigger_id": binding.id.as_str(),
-                "binding_key": binding.binding_key(),
-                "event_id": event.id.0,
-                "reason": reason,
-                "request_id": request_id,
-                "replay_of_event_id": replay_of_event_id,
-            }),
-        )
-        .await
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    async fn append_tier_transition_trust_record(
-        &self,
-        binding: &TriggerBinding,
-        event: &TriggerEvent,
-        replay_of_event_id: Option<&String>,
-        from_tier: AutonomyTier,
-        to_tier: AutonomyTier,
-        reason: &str,
-        request_id: &str,
-    ) -> Result<(), DispatchError> {
-        let mut record = TrustRecord::new(
-            binding.id.as_str().to_string(),
-            "autonomy.tier_transition",
-            Some(DEFAULT_AUTONOMY_BUDGET_REVIEWER.to_string()),
-            TrustOutcome::Denied,
-            event.trace_id.0.clone(),
-            to_tier,
-        );
-        record.metadata.insert(
-            "binding_key".to_string(),
-            serde_json::json!(binding.binding_key()),
-        );
-        record
-            .metadata
-            .insert("event_id".to_string(), serde_json::json!(event.id.0));
-        record.metadata.insert(
-            "from_tier".to_string(),
-            serde_json::json!(from_tier.as_str()),
-        );
-        record
-            .metadata
-            .insert("to_tier".to_string(), serde_json::json!(to_tier.as_str()));
-        record
-            .metadata
-            .insert("reason".to_string(), serde_json::json!(reason));
-        record
-            .metadata
-            .insert("request_id".to_string(), serde_json::json!(request_id));
-        if let Some(replay_of_event_id) = replay_of_event_id {
-            record.metadata.insert(
-                "replay_of_event_id".to_string(),
-                serde_json::json!(replay_of_event_id),
-            );
-        }
-        append_trust_record(&self.event_log, &record)
-            .await
-            .map(|_| ())
-            .map_err(DispatchError::from)
-    }
-
-    async fn append_budget_deferred_event(
-        &self,
-        binding: &TriggerBinding,
-        route: &DispatchUri,
-        event: &TriggerEvent,
-        replay_of_event_id: Option<&String>,
-        reason: &str,
-    ) -> Result<(), DispatchError> {
-        self.append_topic_event(
-            TRIGGER_ATTEMPTS_TOPIC,
-            "budget_deferred",
-            event,
-            Some(binding),
-            None,
-            serde_json::json!({
-                "event_id": event.id.0,
-                "trigger_id": binding.id.as_str(),
-                "binding_key": binding.binding_key(),
-                "handler_kind": route.kind(),
-                "target_uri": route.target_uri(),
-                "reason": reason,
-                "retry_at": next_budget_reset_rfc3339(binding),
-                "replay_of_event_id": replay_of_event_id,
-            }),
-            replay_of_event_id,
-        )
-        .await?;
-        self.append_skipped_outbox_event(
-            binding,
-            route,
-            event,
-            replay_of_event_id,
-            DispatchSkipStage::Predicate,
-            serde_json::json!({
-                "deferred": true,
-                "reason": reason,
-                "retry_at": next_budget_reset_rfc3339(binding),
-            }),
-        )
-        .await
-    }
-
-    async fn append_skipped_outbox_event(
-        &self,
-        binding: &TriggerBinding,
-        route: &DispatchUri,
-        event: &TriggerEvent,
-        replay_of_event_id: Option<&String>,
-        stage: DispatchSkipStage,
-        detail: serde_json::Value,
-    ) -> Result<(), DispatchError> {
-        self.append_topic_event(
-            TRIGGER_OUTBOX_TOPIC,
-            "dispatch_skipped",
-            event,
-            Some(binding),
-            None,
-            serde_json::json!({
-                "event_id": event.id.0,
-                "trigger_id": binding.id.as_str(),
-                "binding_key": binding.binding_key(),
-                "handler_kind": route.kind(),
-                "target_uri": route.target_uri(),
-                "skip_stage": stage.as_str(),
-                "detail": detail,
-                "replay_of_event_id": replay_of_event_id,
-            }),
-            replay_of_event_id,
-        )
-        .await
-    }
-
-    async fn append_topic_event(
-        &self,
-        topic_name: &str,
-        kind: &str,
-        event: &TriggerEvent,
-        binding: Option<&TriggerBinding>,
-        attempt: Option<u32>,
-        payload: serde_json::Value,
-        replay_of_event_id: Option<&String>,
-    ) -> Result<(), DispatchError> {
-        let topic = topic_for_event(event, topic_name)?;
-        let headers = event_headers(event, binding, attempt, replay_of_event_id);
-        self.event_log
-            .append(&topic, LogEvent::new(kind, payload).with_headers(headers))
-            .await
-            .map_err(DispatchError::from)
-            .map(|_| ())
-    }
-
-    async fn emit_action_graph(
-        &self,
-        event: &TriggerEvent,
-        nodes: Vec<RunActionGraphNodeRecord>,
-        edges: Vec<RunActionGraphEdgeRecord>,
-        extra: serde_json::Value,
-    ) -> Result<(), DispatchError> {
-        let mut headers = BTreeMap::new();
-        headers.insert("trace_id".to_string(), event.trace_id.0.clone());
-        headers.insert("event_id".to_string(), event.id.0.clone());
-        let observability = RunObservabilityRecord {
-            schema_version: 1,
-            action_graph_nodes: nodes,
-            action_graph_edges: edges,
-            ..Default::default()
-        };
-        append_action_graph_update(
-            headers,
-            serde_json::json!({
-                "source": "dispatcher",
-                "trace_id": event.trace_id.0,
-                "event_id": event.id.0,
-                "observability": observability,
-                "context": extra,
-            }),
-        )
-        .await
-        .map_err(DispatchError::from)
-    }
-}
-
-async fn dispatch_cancel_requested(
-    event_log: &Arc<AnyEventLog>,
-    binding_key: &str,
-    event_id: &str,
-    replay_of_event_id: Option<&String>,
-) -> Result<bool, DispatchError> {
-    if replay_of_event_id.is_some() {
-        return Ok(false);
-    }
-    let topic = Topic::new(TRIGGER_CANCEL_REQUESTS_TOPIC)
-        .expect("static trigger cancel topic should always be valid");
-    let events = event_log.read_range(&topic, None, usize::MAX).await?;
-    let requested = events
-        .into_iter()
-        .filter(|(_, event)| event.kind == "dispatch_cancel_requested")
-        .filter_map(|(_, event)| {
-            serde_json::from_value::<DispatchCancelRequest>(event.payload).ok()
-        })
-        .collect::<BTreeSet<_>>();
-    Ok(requested
-        .iter()
-        .any(|request| request.binding_key == binding_key && request.event_id == event_id))
-}
-
-async fn sleep_or_cancel_or_request(
-    event_log: &Arc<AnyEventLog>,
-    delay: Duration,
-    binding_key: &str,
-    event_id: &str,
-    replay_of_event_id: Option<&String>,
-    cancel_rx: &mut broadcast::Receiver<()>,
-) -> Result<(), DispatchError> {
-    let sleep = tokio::time::sleep(delay);
-    pin_mut!(sleep);
-    let mut poll = tokio::time::interval(Duration::from_millis(100));
-    loop {
-        tokio::select! {
-            _ = &mut sleep => return Ok(()),
-            _ = recv_cancel(cancel_rx) => {
-                return Err(DispatchError::Cancelled(
-                    "dispatcher shutdown cancelled retry wait".to_string(),
-                ));
-            }
-            _ = poll.tick() => {
-                if dispatch_cancel_requested(event_log, binding_key, event_id, replay_of_event_id).await? {
-                    return Err(DispatchError::Cancelled(
-                        "trigger cancel request cancelled retry wait".to_string(),
-                    ));
-                }
-            }
-        }
-    }
-}
-
-fn build_batched_event(events: Vec<TriggerEvent>) -> Result<TriggerEvent, DispatchError> {
-    let mut iter = events.into_iter();
-    let Some(mut root) = iter.next() else {
-        return Err(DispatchError::Registry(
-            "batch dispatch produced an empty event list".to_string(),
-        ));
-    };
-    let mut batch = Vec::new();
-    batch.push(
-        serde_json::to_value(&root).map_err(|error| DispatchError::Serde(error.to_string()))?,
-    );
-    for event in iter {
-        batch.push(
-            serde_json::to_value(&event)
-                .map_err(|error| DispatchError::Serde(error.to_string()))?,
-        );
-    }
-    root.batch = Some(batch);
-    Ok(root)
-}
-
-fn json_value_to_gate(value: &serde_json::Value) -> String {
-    match value {
-        serde_json::Value::Null => "null".to_string(),
-        serde_json::Value::String(text) => text.clone(),
-        serde_json::Value::Bool(value) => value.to_string(),
-        serde_json::Value::Number(value) => value.to_string(),
-        other => serde_json::to_string(other).unwrap_or_else(|_| "unserializable".to_string()),
-    }
-}
-
-fn event_to_handler_value(event: &TriggerEvent) -> Result<VmValue, DispatchError> {
-    let json =
-        serde_json::to_value(event).map_err(|error| DispatchError::Serde(error.to_string()))?;
-    let value = json_to_vm_value(&json);
-    match (&event.raw_body, value) {
-        (Some(raw_body), VmValue::Dict(dict)) => {
-            let mut map = (*dict).clone();
-            map.insert(
-                "raw_body".to_string(),
-                VmValue::Bytes(Rc::new(raw_body.clone())),
-            );
-            Ok(VmValue::Dict(Rc::new(map)))
-        }
-        (_, other) => Ok(other),
-    }
-}
-
-fn decrement_in_flight(state: &DispatcherRuntimeState) {
-    let previous = state.in_flight.fetch_sub(1, Ordering::Relaxed);
-    if previous == 1 && state.retry_queue_depth.load(Ordering::Relaxed) == 0 {
-        state.idle_notify.notify_waiters();
-    }
-}
-
-fn decrement_retry_queue_depth(state: &DispatcherRuntimeState) {
-    let previous = state.retry_queue_depth.fetch_sub(1, Ordering::Relaxed);
-    if previous == 1 && state.in_flight.load(Ordering::Relaxed) == 0 {
-        state.idle_notify.notify_waiters();
-    }
-}
-
-#[cfg(test)]
-fn install_test_inbox_dequeued_signal(tx: tokio::sync::oneshot::Sender<()>) {
-    TEST_INBOX_DEQUEUED_SIGNAL.with(|slot| {
-        *slot.borrow_mut() = Some(tx);
-    });
-}
-
-#[cfg(not(test))]
-fn notify_test_inbox_dequeued() {}
-
-#[cfg(test)]
-fn notify_test_inbox_dequeued() {
-    TEST_INBOX_DEQUEUED_SIGNAL.with(|slot| {
-        if let Some(tx) = slot.borrow_mut().take() {
-            let _ = tx.send(());
-        }
-    });
-}
-
-pub async fn enqueue_trigger_event<L: EventLog + ?Sized>(
-    event_log: &L,
-    event: &TriggerEvent,
-) -> Result<u64, DispatchError> {
-    let topic = topic_for_event(event, TRIGGER_INBOX_ENVELOPES_TOPIC)?;
-    let headers = event_headers(event, None, None, None);
-    let payload =
-        serde_json::to_value(event).map_err(|error| DispatchError::Serde(error.to_string()))?;
-    event_log
-        .append(
-            &topic,
-            LogEvent::new("event_ingested", payload).with_headers(headers),
-        )
-        .await
-        .map_err(DispatchError::from)
-}
-
-pub fn snapshot_dispatcher_stats() -> DispatcherStatsSnapshot {
-    ACTIVE_DISPATCHER_STATE.with(|slot| {
-        slot.borrow()
-            .as_ref()
-            .map(|state| DispatcherStatsSnapshot {
-                in_flight: state.in_flight.load(Ordering::Relaxed),
-                retry_queue_depth: state.retry_queue_depth.load(Ordering::Relaxed),
-                dlq_depth: state.dlq.lock().expect("dispatcher dlq poisoned").len() as u64,
-            })
-            .unwrap_or_default()
-    })
-}
-
-pub fn clear_dispatcher_state() {
-    ACTIVE_DISPATCHER_STATE.with(|slot| {
-        *slot.borrow_mut() = None;
-    });
-    ACTIVE_DISPATCH_WAIT_LEASE.with(|slot| {
-        *slot.borrow_mut() = None;
-    });
-}
-
-fn dispatch_error_from_vm_error(error: VmError) -> DispatchError {
-    if let Some(wait_id) = crate::stdlib::waitpoint::is_waitpoint_suspension(&error) {
-        return DispatchError::Waiting(format!("waitpoint suspended: {wait_id}"));
-    }
-    if is_cancelled_vm_error(&error) {
-        return DispatchError::Cancelled("dispatcher shutdown cancelled local handler".to_string());
-    }
-    if let VmError::Thrown(VmValue::String(message)) = &error {
-        return DispatchError::Local(message.to_string());
-    }
-    match error_to_category(&error) {
-        ErrorCategory::Timeout => DispatchError::Timeout(error.to_string()),
-        ErrorCategory::ToolRejected => DispatchError::Denied(error.to_string()),
-        ErrorCategory::Cancelled => {
-            DispatchError::Cancelled("dispatcher shutdown cancelled local handler".to_string())
-        }
-        _ => DispatchError::Local(error.to_string()),
-    }
-}
-
-fn dispatch_error_label(error: &DispatchError) -> &'static str {
-    match error {
-        DispatchError::Denied(_) => "denied",
-        DispatchError::Timeout(_) => "timeout",
-        DispatchError::Waiting(_) => "waiting",
-        DispatchError::Cancelled(_) => "cancelled",
-        _ => "failed",
-    }
-}
-
-fn dispatch_success_outcome(route: &DispatchUri, result: &serde_json::Value) -> &'static str {
-    match route {
-        DispatchUri::Worker { .. } => "enqueued",
-        DispatchUri::Persona { .. } => "recorded",
-        DispatchUri::A2a { .. }
-            if result.get("kind").and_then(|value| value.as_str()) == Some("a2a_task_handle") =>
-        {
-            "pending"
-        }
-        DispatchUri::A2a { .. } => "completed",
-        DispatchUri::Local { .. } => "success",
-    }
-}
-
-fn dispatch_node_id(
-    route: &DispatchUri,
-    binding_key: &str,
-    event_id: &str,
-    attempt: u32,
-) -> String {
-    let prefix = match route {
-        DispatchUri::A2a { .. } => "a2a",
-        _ => "dispatch",
-    };
-    format!("{prefix}:{binding_key}:{event_id}:{attempt}")
-}
-
-fn dispatch_node_kind(route: &DispatchUri) -> &'static str {
-    match route {
-        DispatchUri::A2a { .. } => ACTION_GRAPH_NODE_KIND_A2A_HOP,
-        DispatchUri::Worker { .. } => ACTION_GRAPH_NODE_KIND_WORKER_ENQUEUE,
-        _ => ACTION_GRAPH_NODE_KIND_DISPATCH,
-    }
-}
-
-fn dispatch_node_label(route: &DispatchUri) -> String {
-    match route {
-        DispatchUri::A2a { target, .. } => crate::a2a::target_agent_label(target),
-        _ => route.target_uri(),
-    }
-}
-
-fn dispatch_target_agent(route: &DispatchUri) -> Option<String> {
-    match route {
-        DispatchUri::A2a { target, .. } => Some(crate::a2a::target_agent_label(target)),
-        _ => None,
-    }
-}
-
-fn dispatch_entry_edge_kind(route: &DispatchUri, has_predicate: bool) -> &'static str {
-    match route {
-        DispatchUri::A2a { .. } => ACTION_GRAPH_EDGE_KIND_A2A_DISPATCH,
-        _ if has_predicate => ACTION_GRAPH_EDGE_KIND_PREDICATE_GATE,
-        _ => ACTION_GRAPH_EDGE_KIND_TRIGGER_DISPATCH,
-    }
-}
-
-fn signature_status_label(status: &crate::triggers::SignatureStatus) -> &'static str {
-    match status {
-        crate::triggers::SignatureStatus::Verified => "verified",
-        crate::triggers::SignatureStatus::Unsigned => "unsigned",
-        crate::triggers::SignatureStatus::Failed { .. } => "failed",
-    }
-}
-
-fn trigger_node_metadata(event: &TriggerEvent) -> BTreeMap<String, serde_json::Value> {
-    let mut metadata = BTreeMap::new();
-    metadata.insert(
-        "provider".to_string(),
-        serde_json::json!(event.provider.as_str()),
-    );
-    metadata.insert("event_kind".to_string(), serde_json::json!(event.kind));
-    metadata.insert(
-        "dedupe_key".to_string(),
-        serde_json::json!(event.dedupe_key),
-    );
-    metadata.insert(
-        "signature_status".to_string(),
-        serde_json::json!(signature_status_label(&event.signature_status)),
-    );
-    metadata
-}
-
-fn trigger_event_persona_metadata(event: &TriggerEvent) -> BTreeMap<String, String> {
-    let mut metadata = BTreeMap::new();
-    metadata.insert("trigger_event_id".to_string(), event.id.0.clone());
-    metadata.insert("event_id".to_string(), event.id.0.clone());
-    metadata.insert("dedupe_key".to_string(), event.dedupe_key.clone());
-    metadata.insert("trace_id".to_string(), event.trace_id.0.clone());
-    if let Some(tenant_id) = &event.tenant_id {
-        metadata.insert("tenant_id".to_string(), tenant_id.0.clone());
-    }
-    for (key, value) in &event.headers {
-        metadata.insert(format!("header.{key}"), value.clone());
-    }
-    if let Ok(payload) = serde_json::to_value(&event.provider_payload) {
-        collect_persona_payload_metadata("", &payload, &mut metadata);
-        if let Some(raw) = payload.get("raw") {
-            collect_persona_payload_metadata("", raw, &mut metadata);
-        }
-    }
-    metadata
-}
-
-fn collect_persona_payload_metadata(
-    prefix: &str,
-    value: &serde_json::Value,
-    metadata: &mut BTreeMap<String, String>,
-) {
-    let serde_json::Value::Object(object) = value else {
-        return;
-    };
-    for (key, value) in object {
-        let path = if prefix.is_empty() {
-            key.clone()
-        } else {
-            format!("{prefix}.{key}")
-        };
-        match value {
-            serde_json::Value::String(text) => {
-                metadata.insert(path, text.clone());
-            }
-            serde_json::Value::Number(number) => {
-                metadata.insert(path, number.to_string());
-            }
-            serde_json::Value::Bool(flag) => {
-                metadata.insert(path, flag.to_string());
-            }
-            serde_json::Value::Object(_) if prefix.is_empty() => {
-                collect_persona_payload_metadata(&path, value, metadata);
-            }
-            _ => {}
-        }
-    }
-}
-
-fn dispatch_node_metadata(
-    route: &DispatchUri,
-    binding: &TriggerBinding,
-    event: &TriggerEvent,
-    attempt: u32,
-) -> BTreeMap<String, serde_json::Value> {
-    let mut metadata = BTreeMap::new();
-    metadata.insert("handler_kind".to_string(), serde_json::json!(route.kind()));
-    metadata.insert(
-        "target_uri".to_string(),
-        serde_json::json!(route.target_uri()),
-    );
-    metadata.insert("attempt".to_string(), serde_json::json!(attempt));
-    metadata.insert(
-        "trigger_id".to_string(),
-        serde_json::json!(binding.id.as_str()),
-    );
-    metadata.insert("event_id".to_string(), serde_json::json!(event.id.0));
-    if let Some(target_agent) = dispatch_target_agent(route) {
-        metadata.insert("target_agent".to_string(), serde_json::json!(target_agent));
-    }
-    if let DispatchUri::Worker { queue } = route {
-        metadata.insert("queue_name".to_string(), serde_json::json!(queue));
-    }
-    if let DispatchUri::Persona { name } = route {
-        metadata.insert("persona".to_string(), serde_json::json!(name));
-    }
-    metadata
-}
-
-fn dispatch_success_metadata(
-    route: &DispatchUri,
-    binding: &TriggerBinding,
-    event: &TriggerEvent,
-    attempt: u32,
-    result: &serde_json::Value,
-) -> BTreeMap<String, serde_json::Value> {
-    let mut metadata = dispatch_node_metadata(route, binding, event, attempt);
-    match route {
-        DispatchUri::A2a { .. } => {
-            if let Some(task_id) = result
-                .get("task_id")
-                .or_else(|| result.get("id"))
-                .and_then(|value| value.as_str())
-            {
-                metadata.insert("task_id".to_string(), serde_json::json!(task_id));
-            }
-            if let Some(state) = result.get("state").and_then(|value| value.as_str()) {
-                metadata.insert("state".to_string(), serde_json::json!(state));
-            }
-        }
-        DispatchUri::Worker { .. } => {
-            if let Some(job_event_id) = result.get("job_event_id").and_then(|value| value.as_u64())
-            {
-                metadata.insert("job_event_id".to_string(), serde_json::json!(job_event_id));
-            }
-            if let Some(response_topic) = result
-                .get("response_topic")
-                .and_then(|value| value.as_str())
-            {
-                metadata.insert(
-                    "response_topic".to_string(),
-                    serde_json::json!(response_topic),
-                );
-            }
-        }
-        DispatchUri::Persona { .. } => {
-            if let Some(receipt_id) = result.get("receipt_id").and_then(|value| value.as_str()) {
-                metadata.insert("receipt_id".to_string(), serde_json::json!(receipt_id));
-            }
-            if let Some(status) = result.get("status").and_then(|value| value.as_str()) {
-                metadata.insert("status".to_string(), serde_json::json!(status));
-            }
-        }
-        DispatchUri::Local { .. } => {}
-    }
-    metadata
-}
-
-fn dispatch_error_metadata(
-    route: &DispatchUri,
-    binding: &TriggerBinding,
-    event: &TriggerEvent,
-    attempt: u32,
-    error: &DispatchError,
-) -> BTreeMap<String, serde_json::Value> {
-    let mut metadata = dispatch_node_metadata(route, binding, event, attempt);
-    metadata.insert("error".to_string(), serde_json::json!(error.to_string()));
-    metadata
-}
-
-fn retry_node_metadata(
-    binding: &TriggerBinding,
-    event: &TriggerEvent,
-    attempt: u32,
-    delay: Duration,
-    error: &DispatchError,
-) -> BTreeMap<String, serde_json::Value> {
-    let mut metadata = BTreeMap::new();
-    metadata.insert(
-        "trigger_id".to_string(),
-        serde_json::json!(binding.id.as_str()),
-    );
-    metadata.insert("event_id".to_string(), serde_json::json!(event.id.0));
-    metadata.insert("attempt".to_string(), serde_json::json!(attempt));
-    metadata.insert("delay_ms".to_string(), serde_json::json!(delay.as_millis()));
-    metadata.insert("error".to_string(), serde_json::json!(error.to_string()));
-    metadata
-}
-
-fn split_binding_key(binding_key: &str) -> (String, u32) {
-    let Some((binding_id, suffix)) = binding_key.rsplit_once("@v") else {
-        return (binding_key.to_string(), 0);
-    };
-    let version = suffix.parse::<u32>().unwrap_or(0);
-    (binding_id.to_string(), version)
-}
-
-fn binding_key_from_parts(trigger_id: &str, binding_version: Option<u32>) -> String {
-    match binding_version {
-        Some(version) => format!("{trigger_id}@v{version}"),
-        None => trigger_id.to_string(),
-    }
-}
-
-fn tenant_id(event: &TriggerEvent) -> Option<&str> {
-    event.tenant_id.as_ref().map(|tenant| tenant.0.as_str())
-}
-
-fn current_unix_ms() -> i64 {
-    unix_ms(crate::triggers::test_util::clock::now_utc())
-}
-
-fn unix_ms(timestamp: time::OffsetDateTime) -> i64 {
-    (timestamp.unix_timestamp_nanos() / 1_000_000) as i64
-}
-
-fn accepted_at_ms(headers: Option<&BTreeMap<String, String>>, event: &TriggerEvent) -> i64 {
-    lifecycle_header_ms(headers, TRIGGER_ACCEPTED_AT_MS_HEADER)
-        .unwrap_or_else(|| unix_ms(event.received_at))
-}
-
-fn queue_appended_at_ms(headers: Option<&BTreeMap<String, String>>, event: &TriggerEvent) -> i64 {
-    lifecycle_header_ms(headers, TRIGGER_QUEUE_APPENDED_AT_MS_HEADER)
-        .unwrap_or_else(|| accepted_at_ms(headers, event))
-}
-
-fn lifecycle_header_ms(headers: Option<&BTreeMap<String, String>>, name: &str) -> Option<i64> {
-    headers
-        .and_then(|headers| headers.get(name))
-        .and_then(|value| value.parse::<i64>().ok())
-}
-
-fn duration_between_ms(later_ms: i64, earlier_ms: i64) -> Duration {
-    Duration::from_millis(later_ms.saturating_sub(earlier_ms).max(0) as u64)
-}
-
-fn dispatch_result_status(result: &Result<serde_json::Value, DispatchError>) -> &'static str {
-    match result {
-        Ok(_) => "succeeded",
-        Err(DispatchError::Waiting(_)) => "waiting",
-        Err(DispatchError::Cancelled(_)) => "cancelled",
-        Err(DispatchError::Denied(_)) => "denied",
-        Err(DispatchError::Timeout(_)) => "timeout",
-        Err(_) => "failed",
-    }
-}
-
-fn is_cancelled_vm_error(error: &VmError) -> bool {
-    matches!(
-        error,
-        VmError::Thrown(VmValue::String(message))
-            if message.starts_with("kind:cancelled:")
-    ) || matches!(error_to_category(error), ErrorCategory::Cancelled)
-}
-
-fn event_headers(
-    event: &TriggerEvent,
-    binding: Option<&TriggerBinding>,
-    attempt: Option<u32>,
-    replay_of_event_id: Option<&String>,
-) -> BTreeMap<String, String> {
-    let mut headers = BTreeMap::new();
-    headers.insert("event_id".to_string(), event.id.0.clone());
-    headers.insert("trace_id".to_string(), event.trace_id.0.clone());
-    headers.insert("provider".to_string(), event.provider.as_str().to_string());
-    headers.insert("kind".to_string(), event.kind.clone());
-    if let Some(replay_of_event_id) = replay_of_event_id {
-        headers.insert("replay_of_event_id".to_string(), replay_of_event_id.clone());
-    }
-    if let Some(tenant_id) = event.tenant_id.as_ref() {
-        headers.insert("tenant_id".to_string(), tenant_id.0.clone());
-    }
-    if let Some(binding) = binding {
-        headers.insert("trigger_id".to_string(), binding.id.as_str().to_string());
-        headers.insert("binding_key".to_string(), binding.binding_key());
-        headers.insert(
-            "handler_kind".to_string(),
-            DispatchUri::from(&binding.handler).kind().to_string(),
-        );
-    }
-    if let Some(attempt) = attempt {
-        headers.insert("attempt".to_string(), attempt.to_string());
-    }
-    headers
-}
-
-fn topic_for_event(event: &TriggerEvent, topic_name: &str) -> Result<Topic, DispatchError> {
-    let topic = Topic::new(topic_name)
-        .expect("static trigger dispatcher topic names should always be valid");
-    match event.tenant_id.as_ref() {
-        Some(tenant_id) => crate::tenant_topic(tenant_id, &topic).map_err(DispatchError::from),
-        None => Ok(topic),
-    }
-}
-
-fn worker_queue_priority(
-    binding: &super::registry::TriggerBinding,
-    event: &TriggerEvent,
-) -> crate::WorkerQueuePriority {
-    match event
-        .headers
-        .get("priority")
-        .map(|value| value.trim().to_ascii_lowercase())
-        .as_deref()
-    {
-        Some("high") => crate::WorkerQueuePriority::High,
-        Some("low") => crate::WorkerQueuePriority::Low,
-        _ => binding.dispatch_priority,
-    }
-}
-
-const TEST_FAIL_BEFORE_OUTBOX_ENV: &str = "HARN_TEST_DISPATCHER_FAIL_BEFORE_OUTBOX";
-
-fn maybe_fail_before_outbox() {
-    if std::env::var_os(TEST_FAIL_BEFORE_OUTBOX_ENV).is_some() {
-        std::process::exit(86);
-    }
-}
-
-fn now_rfc3339() -> String {
-    crate::triggers::test_util::clock::now_utc()
-        .format(&time::format_description::well_known::Rfc3339)
-        .unwrap_or_else(|_| "1970-01-01T00:00:00Z".to_string())
-}
-
-fn next_budget_reset_rfc3339(binding: &TriggerBinding) -> String {
-    let now = crate::triggers::test_util::clock::now_utc();
-    let reset = if binding.hourly_cost_usd.is_some() {
-        let next_hour = (now.unix_timestamp() / 3_600 + 1) * 3_600;
-        time::OffsetDateTime::from_unix_timestamp(next_hour).unwrap_or(now)
-    } else {
-        let next_day = ((now.unix_timestamp() / 86_400) + 1) * 86_400;
-        time::OffsetDateTime::from_unix_timestamp(next_day).unwrap_or(now)
-    };
-    reset
-        .format(&time::format_description::well_known::Rfc3339)
-        .unwrap_or_else(|_| "1970-01-01T00:00:00Z".to_string())
-}
-
-fn now_unix_ms() -> i64 {
-    (crate::triggers::test_util::clock::now_utc().unix_timestamp_nanos() / 1_000_000) as i64
-}
-
-fn cancelled_dispatch_outcome(
-    binding: &TriggerBinding,
-    route: &DispatchUri,
-    event: &TriggerEvent,
-    replay_of_event_id: Option<String>,
-    attempt_count: u32,
-    error: String,
-) -> DispatchOutcome {
-    DispatchOutcome {
-        trigger_id: binding.id.as_str().to_string(),
-        binding_key: binding.binding_key(),
-        event_id: event.id.0.clone(),
-        attempt_count,
-        status: DispatchStatus::Cancelled,
-        handler_kind: route.kind().to_string(),
-        target_uri: route.target_uri(),
-        replay_of_event_id,
-        result: None,
-        error: Some(error),
-    }
-}
-
-async fn recv_cancel(cancel_rx: &mut broadcast::Receiver<()>) {
-    let _ = cancel_rx.recv().await;
 }
 
 #[cfg(test)]
